- Add TPC-H Data as test dependency (automatically downloads from s3 on first execution) - Update avg function to divide using double instead of integers - Update classpath file system to return correct size information (to allow cp use of Parquet files) - Enhance BaseTestQuery to support all query types (sql, physical logical) and execution from file - Add TPC-H queries updated to use TPC-H test data (disabled) - Update field name matching to be case insensitive. - Treat CHAR literals like VARCHAR in Optiq - Add clear as part of VectorUtil - Add hash aggregation physical plan executions using tpch data.

+++ exec/java-exec/src/main/codegen/templates/AggrTypeFunctions2.java
@@ -88,7 +88,7 @@ public static class ${type.inputType}${aggrtype.className} implements DrillAggFu
 
   @Override
   public void output() {
-    out.value = sum.value/count.value;
+    out.value = sum.value / ((double) count.value);
   }
 
   @Override
